+++ uefi-raw/CHANGELOG.md
@@ -9,6 +9,7 @@
   `MacAddress` to streamline the API with `core::net`.
 - Added `::into_core_addr()` for `IpAddress`
 - Added `::into_ethernet_addr()` for `MacAddress`
+- Added `::ZERO` constant for `IpAddress`
 
 ## Changed
 - **Breaking:** The MSRV is now 1.85.1 and the crate uses the Rust 2024 edition.